utils/requirements.txt
python-dotenv>=1.0.0
azure-storage-blob>=12.0.0
azure-identity>=1.15.0
